Divide logic file into separated files
